rename package tools.vitruv.framework.change.models to .propagation
