+++ bundles/framework/tools.vitruv.framework.change/src/tools/vitruv/framework/change/propagation/ChangePropagationAbortCause.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.change.models;
+package tools.vitruv.framework.change.propagation;
 
 /**
  * Marker interface for all classes, which could become a cause for a
