+++ bundles/framework/tools.vitruv.framework.views/src/tools/vitruv/framework/views/impl/ChangeRecordingView.xtend
@@ -9,8 +9,8 @@ import org.eclipse.emf.ecore.resource.Resource
 import org.eclipse.emf.ecore.resource.ResourceSet
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl
 import org.eclipse.xtend.lib.annotations.Accessors
-import tools.vitruv.framework.change.models.ChangePropagationAbortCause
-import tools.vitruv.framework.change.models.ChangePropagationListener
+import tools.vitruv.framework.change.propagation.ChangePropagationAbortCause
+import tools.vitruv.framework.change.propagation.ChangePropagationListener
 import tools.vitruv.framework.change.recording.ChangeRecorder
 import tools.vitruv.framework.views.ChangeableViewSource
 import tools.vitruv.framework.views.View
